Spliting code and adding comments
