+++ src/App.tsx
@@ -1,12 +1,5 @@
 import "./App.css";
-import {
-  Meeting,
-  MeetingSeries,
-  TalkingPoint,
-  TalkingPointTicket,
-} from "./types";
-import { useEffect, useState } from "react";
-import { getSeries } from "./server";
+import { TalkingPoint } from "./types";
 import { Header } from "./components/Header";
 import { TalkingPointForm } from "./components/TalkingPointForm";
 import { TalkingPointTickets } from "./components/TalkingPointTickets";
@@ -15,85 +8,18 @@ import { Draggable } from "./components/Draggable";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import { DndProvider } from "react-dnd";
 import { Droppable } from "./components/Droppable";
+import { useMeeting } from "./components/useMeeting";
 
 function App() {
-  const [series, setSeries] = useState<MeetingSeries>();
-  const [currentMeeting] = series?.meetings || [];
-
-  useEffect(() => {
-    getSeries().then(setSeries);
-  }, []);
-
-  const updateMeeting = (updatedMeeting: Meeting) => {
-    const meetings = series!.meetings!.map((meeting) =>
-      meeting.id === meeting.id ? updatedMeeting : meeting
-    );
-
-    setSeries({ ...series!, meetings });
-  };
-
-  const updateTalkingPoint = (talkingPoint: TalkingPoint) => {
-    const agendaItems = currentMeeting.agendaItems.map((item) =>
-      talkingPoint.id === item.id ? talkingPoint : item
-    );
-
-    updateMeeting({
-      ...currentMeeting,
-      agendaItems,
-    });
-  };
-
-  const updateTalkingPointGeneralData = (talkingPoint: TalkingPoint) => {
-    const { linkedTickets, tickets } = currentMeeting.agendaItems.find(
-      (item) => talkingPoint.id === item.id
-    ) as TalkingPoint;
-
-    updateTalkingPoint({
-      ...talkingPoint,
-      linkedTickets,
-      tickets,
-    });
-  };
-
-  const linkTicket = (
-    talkingPoint: TalkingPoint,
-    ticket: TalkingPointTicket
-  ) => {
-    updateTalkingPoint({
-      ...talkingPoint,
-      linkedTickets: [...talkingPoint.linkedTickets, ticket.id],
-      tickets: [...talkingPoint.tickets, ticket],
-    });
-  };
-
-  const unlinkTicket = (
-    talkingPoint: TalkingPoint,
-    ticket: TalkingPointTicket
-  ) => {
-    updateTalkingPoint({
-      ...talkingPoint,
-      linkedTickets: talkingPoint.linkedTickets.filter(
-        (id) => id !== ticket.id
-      ),
-      tickets: talkingPoint.tickets.filter(({ id }) => id !== ticket.id),
-    });
-  };
-
-  const moveTalkingPoint = (
-    talkingPoint: TalkingPoint,
-    newPosition: number
-  ) => {
-    const agendaItems = currentMeeting.agendaItems.filter(
-      (item) => talkingPoint.id !== item.id
-    );
-
-    agendaItems.splice(newPosition, 0, talkingPoint);
-
-    updateMeeting({
-      ...currentMeeting,
-      agendaItems,
-    });
-  };
+  // We could break this into useSeries and useMeetings({series}), but I don't think it is necessary in a POC
+  const {
+    series,
+    currentMeeting,
+    moveTalkingPoint,
+    updateTalkingPointGeneralData,
+    linkTicket,
+    unlinkTicket,
+  } = useMeeting();
 
   if (!series) return <span>This is loading</span>;
 
@@ -103,31 +29,30 @@ function App() {
         <Header title={series.name} date={series.date} onEdit={() => {}} />
         <h3>Agenda Items</h3>
         {currentMeeting.agendaItems.map((item, index) => (
-          <div key={item.id}>
-            <Droppable type="item" index={index}>
-              <Draggable
-                type="item"
+          // I won't bother preparing the same for sections, you get the idea.
+          // I am using id as keys because they are stable and now we have this during creation
+          <Droppable type="item" key={item.id} index={index}>
+            <Draggable
+              type="item"
+              talkingPoint={item as TalkingPoint}
+              onMove={moveTalkingPoint}
+            >
+              <TalkingPointForm
                 talkingPoint={item as TalkingPoint}
-                onMove={moveTalkingPoint}
+                onSubmit={updateTalkingPointGeneralData}
               >
-                <TalkingPointForm
-                  talkingPoint={item as TalkingPoint}
-                  onSubmit={updateTalkingPointGeneralData}
-                  onMove={moveTalkingPoint}
-                >
-                  <TalkingPointTickets
-                    tickets={(item as TalkingPoint).tickets}
-                    onLinkTicket={(ticket) =>
-                      linkTicket(item as TalkingPoint, ticket)
-                    }
-                    onUnlinkTicket={(ticket) =>
-                      unlinkTicket(item as TalkingPoint, ticket)
-                    }
-                  />
-                </TalkingPointForm>
-              </Draggable>
-            </Droppable>
-          </div>
+                <TalkingPointTickets // breaking tickets into their own component without dependencies from talking point form. This way we don't need to nest that much
+                  tickets={(item as TalkingPoint).tickets}
+                  onLinkTicket={(ticket) =>
+                    linkTicket(item as TalkingPoint, ticket)
+                  }
+                  onUnlinkTicket={(ticket) =>
+                    unlinkTicket(item as TalkingPoint, ticket)
+                  }
+                />
+              </TalkingPointForm>
+            </Draggable>
+          </Droppable>
         ))}
       </div>
     </DndProvider>
